users save/ delete

+++ apetit/models/restaurant.js
@@ -2,20 +2,20 @@ const router      = require('express').Router();
 const db = require('../db/config');
 
 const Restaurants ={
-	findAll: ()=> db.manyOrNone('SELECT * FROM restaurants'), 
+	findAllByUser: (userId)=> db.manyOrNone('SELECT * FROM restaurants WHERE user_id=$1', [userId]), 
 
-	findById: (id)=>db.one(`SELECT * FROM restaurants WHERE  id= $1`, [id]), 
+	findById: (id, userId)=>db.one(`SELECT * FROM restaurants WHERE  id= $1 AND user_id=$2`, [id, userId]), 
 
 
-  create: (name, location, latitude, longitude, averagecost, pricerange, thunmpic, cuisines, ratingcolor, aggregaterating) => {
+  create: (name, location, latitude, longitude, averagecost, pricerange, thunmpic, cuisines, ratingcolor, aggregaterating, userId) => {
     return db.one(
-  		`INSERT INTO restaurants (name, location, latitude, longitude, averagecost, pricerange, thunmpic, cuisines, ratingcolor, aggregaterating) VALUES ($1, $2, $3, $4, $5, $6, $7, $8, $9, $10) returning id`,
-      [name, location, latitude, longitude, averagecost, pricerange, thunmpic, cuisines, ratingcolor, aggregaterating]
+  		`INSERT INTO restaurants (name, location, latitude, longitude, averagecost, pricerange, thunmpic, cuisines, ratingcolor, aggregaterating,user_id) VALUES ($1, $2, $3, $4, $5, $6, $7, $8, $9, $10, $11) returning *`,
+      [name, location, latitude, longitude, averagecost, pricerange, thunmpic, cuisines, ratingcolor, aggregaterating, userId]
   	);
   },
 
 
-	delete: (id) => db.none('DELETE FROM restaurants WHERE id = $1', [id])
+	delete: (id, userId) => db.none('DELETE FROM restaurants WHERE id = $1 AND user_id=$2', [id, userId])
 };
